requirements.txt
requests==2.9.1
python-docx==0.8.6
coverage==4.4.2
ddt==1.1.0
git+https://github.com/elifesciences/elife-tools.git@df537e13da03accf1f4dee83c6c0892742860b07#egg=elifetools
configparser==3.5.0
git+https://github.com/Medium/medium-sdk-python.git@bdf34becf6dd24e3b1fc1048c36416bcba9fe0d6#egg=medium
mock==1.3.0
ddt==1.1.0
python-slugify==1.2.4
